+++ .gitignore
@@ -1,0 +1,4 @@
+.DS_Store
+queread/
+Pipfile
+Pipfile.lock
