+++ README.md
@@ -1,0 +1,7 @@
+# quereadai
+QuereadAI
+
+Daeyong Kwon
+Andrew DiBiasio
+Ethan Saddler
+Ray Wang
