+++ requirements.txt
@@ -1,0 +1,113 @@
+aiofiles==23.2.1
+aiohttp==3.8.6
+aiosignal==1.3.1
+altair==5.1.2
+annotated-types==0.6.0
+anyio==3.7.1
+async-timeout==4.0.3
+attrs==23.1.0
+Authlib==1.2.1
+backoff==2.2.1
+bcrypt==4.0.1
+blinker==1.6.3
+certifi==2023.7.22
+cffi==1.16.0
+charset-normalizer==3.3.0
+chroma-hnswlib==0.7.3
+chromadb==0.4.14
+click==8.1.7
+coloredlogs==15.0.1
+contourpy==1.1.1
+cryptography==41.0.4
+cycler==0.12.1
+dataclasses-json==0.6.1
+exceptiongroup==1.1.3
+fastapi==0.103.2
+ffmpy==0.3.1
+filelock==3.12.4
+Flask==3.0.0
+flatbuffers==23.5.26
+fonttools==4.43.1
+frozenlist==1.4.0
+fsspec==2023.9.2
+gradio==3.47.1
+gradio_client==0.6.0
+graphlib-backport==1.0.3
+greenlet==3.0.0
+grpcio==1.59.0
+h11==0.14.0
+httpcore==0.18.0
+httptools==0.6.0
+httpx==0.25.0
+huggingface-hub==0.17.3
+humanfriendly==10.0
+idna==3.4
+importlib-metadata==6.8.0
+importlib-resources==6.1.0
+itsdangerous==2.1.2
+Jinja2==3.1.2
+jsonpatch==1.33
+jsonpointer==2.4
+jsonschema==4.19.1
+jsonschema-specifications==2023.7.1
+kiwisolver==1.4.5
+langchain==0.0.314
+langsmith==0.0.43
+MarkupSafe==2.1.3
+marshmallow==3.20.1
+matplotlib==3.7.3
+monotonic==1.6
+mpmath==1.3.0
+multidict==6.0.4
+mypy-extensions==1.0.0
+numpy==1.24.4
+onnxruntime==1.16.1
+openai==0.28.1
+orjson==3.9.9
+overrides==7.4.0
+packaging==23.2
+pandas==2.0.3
+Pillow==10.0.1
+pkgutil_resolve_name==1.3.10
+posthog==3.0.2
+protobuf==4.24.4
+pulsar-client==3.3.0
+pycparser==2.21
+pydantic==2.4.2
+pydantic_core==2.10.1
+pydub==0.25.1
+pyparsing==3.1.1
+pypdf==3.16.4
+PyPika==0.48.9
+python-dateutil==2.8.2
+python-dotenv==1.0.0
+python-multipart==0.0.6
+pytz==2023.3.post1
+PyYAML==6.0.1
+referencing==0.30.2
+regex==2023.10.3
+requests==2.31.0
+rpds-py==0.10.6
+semantic-version==2.10.0
+six==1.16.0
+sniffio==1.3.0
+SQLAlchemy==2.0.22
+starlette==0.27.0
+sympy==1.12
+tenacity==8.2.3
+tiktoken==0.5.1
+tokenizers==0.14.1
+toolz==0.12.0
+tqdm==4.66.1
+typer==0.9.0
+typing-inspect==0.9.0
+typing_extensions==4.8.0
+tzdata==2023.3
+urllib3==2.0.6
+uvicorn==0.23.2
+uvloop==0.18.0
+watchfiles==0.21.0
+websockets==11.0.3
+Werkzeug==3.0.0
+yarl==1.9.2
+zipp==3.17.0
